requirements.txt
amqp==2.3.2
aniso8601==4.0.1
appnope==0.1.0
asn1crypto==0.24.0
Babel==2.6.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.4
celery==4.2.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
configparser==3.5.0
cryptography==2.3.1
decorator==4.3.0
Deprecated==1.2.3
dnspython==1.15.0
enum34==1.1.6
eventlet==0.24.1
Flask==1.0.2
Flask-RESTful==0.3.6
Flask-Session==0.3.1
Flask-SocketIO==3.0.2
flower==0.9.2
futures==3.1.1
greenlet==0.4.15
gunicorn==19.9.0
idna==2.7
ipaddress==1.0.22
ipdb==0.11
ipython==5.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
monotonic==1.5
oauthlib==2.1.0
parso==0.3.1
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycparser==2.19
PyGithub==1.43.3
Pygments==2.2.0
PyJWT==1.6.4
PyMySQL==0.9.2
python-dotenv==0.9.1
python-engineio==2.3.2
python-socketio==2.0.0
pytz==2018.7
requests==2.20.0
requests-oauthlib==1.0.0
scandir==1.9.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.12
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
